requirements.txt
ckanapi==2.2
get-cached==0.0.1
